+++ CHANGES.md
@@ -1,2 +1,3 @@
 [unreleased]
 * initial pass
+* rename to `test-reports`